Relationship Between Pointers and Arrays
Outputting elements of an array using cout directly with the array in question and getting the same output by using a pointer variable.  Since an array name is itself a pointer to the first element we can use it instead of declaring a unique pointer variable directed to the array elements.  Sounds kinda Linked List-y (...Linked List-ish?....  Linked List-like?.... whatev's)